ADD content negotiation to controller classes
